Fix for failure in fatal error (OOM) test
On v6.9.1 we get a second OOM in v8 if we call the stacktrace API
in the fatal error (OOM) scenario. It worked OK on v6.3.0. Fix
is to remove the call for now. This means we won't provide a JS
stacktrace in the NodeReport for this particular scenario. The v8
stderr output does show a single JS frame, so we may be able to
obtain that for nodereport via another route at some point.