Opt-in hooks by default
This changes the default action on require('nodereport') to opt-in
to the exception and fatal-error hooks, and to enable the signal
handler. So 'node -r nodereport application.js' will get all the
functionality without any code changes.

For users programming to the API, a separate api-only entry point,
require('nodereport/api') is provided. This does not enable the
hooks and signal unless and until .setEvents() is called.

Also includes a change to the behaviour on uncaught exception:
default behaviour is now the same as the node default, ie no
SIGILL/core dump