Fix behaviour on exception to match node default
This patch switches nodereport behaviour on uncaught exceptions
to match node default behaviour. Previously nodereport was
triggering an abort/core dump, unless the .setCoreDump() API or
the NODEREPORT_COREDUMP env var was used to switch that off.

We now detect whether the user specified the
-abort-on-uncaught-exception option and honor that.

The .setCoreDump() API and NODEREPORT_COREDUMP env var are
redundant and are removed. This will be a semver major change.

PR-URL: #36
Reviewed-By: George Adams <george.adams@uk.ibm.com>
Reviewed-By: Richard Lau <riclau@uk.ibm.com>