Add the list of library files loaded by the process to nodereport.
This adds an extra section to the System Information nodereport provides
that shows the libraries loaded by the application. This allows the user
to identify any native libraries loaded by node modules as well as any
other system libraries that have been loaded by node.

It provides the information reported by the operating system. This is
similar to the output of ldd however it includes libraries loaded at
runtime rather than just those that the executable is linked against.

PR-URL: #38
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>
Reviewed-By: Richard Lau <riclau@uk.ibm.com>