Rename nodereport module to node-report
Provides code, documentation, test and demo changes required
for the migration of the nodereport project to node-report. The
main changes are in the build, tests and demos that use the module
name, and in the name of the report file. Remaining changes are
mostly in comments. Internal variable and macro names are not
changed.