Provide getReport API to return the contents of node-report.
These changes add a new function to the node-report API - getReport(). This
returns the string contents of node report to the caller instead of writing
it to a file and returning the file name. This makes it easier for other tools
to build on node-report.

To implement this I switched from using fprintf for output to C++ streams. This
makes it trivial to switch between using a file stream or a string stream for
output and the getReport path just passing a different stream object to the
triggerReport path.

PR-URL: #48
Reviewed-By: Richard Chamberlain <richard_chamberlain@uk.ibm.com>
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>
Reviewed-By: Richard Lau <riclau@uk.ibm.com>