Use nan helpers to set exports
Fixes compilation errors when built on Node.js master with V8 7.4.

`NAN_EXPORT()` requires the function name exported to JavaScript land
and the exported C++ method to match. Change the exported C++ methods
to begin with a lower case character to preserve the existing casing of
the name exported to JavaScript.

PR-URL: https://github.com/nodejs/node-report/pull/125
Fixes: https://github.com/nodejs/node-report/issues/116
Reviewed-By: Sam Roberts <vieuxtech@gmail.com>