Restore uppercase C++ method names
fabac78 changed the exported C++ method names to start with a lower case
character to match the name exported to JavaScript land. This turns out
to be unnecessary if `Nan::SetMethod()` is used instead of
`NAN_EXPORT()`. Restore the original casing of the C++ methods to be
consistent with the non-exported C++ methods.

PR-URL: https://github.com/nodejs/node-report/pull/127
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>
Reviewed-By: Sam Roberts <vieuxtech@gmail.com>