build: disable compiler inlining on z/OS
In Node v14 on z/OS, test `test-exception.js` fails due to a
compiler issue with inlining.

This PR works around it by disabling compiler inlining.

I confirmed that, with this change, `npm test` continues to pass in
Node v8.17.0 (which uses D190508 compiler version), as well as in
Node v12.16.1 (which uses D191122 as in v14).

Once the new Wyvern on z compiler is available for use in Node v14,
we'll revisit and undo this change if Woz resolves the issue.